INFUND-5199 - Review fixes.
Former-commit-id: 29bbf39d698ec084bed77a32b6138b5e076f8813

+++ ifs-data-service/src/test/java/com/worth/ifs/competition/controller/MilestoneControllerIntegrationTest.java
@@ -76,7 +76,7 @@ public class MilestoneControllerIntegrationTest extends BaseControllerIntegratio
 
     @Test
     public void testCreateSingleMilestone() throws Exception {
-        Competition newCompetition  = competitionRepository.save(newCompetition().withId(null).build());
+        Competition newCompetition  = competitionRepository.save(newCompetition().withId((Long)null).build());
 
         List<MilestoneResource> milestones = getMilestonesForCompetition(newCompetition.getId());
         assertNotNull(milestones);
@@ -91,7 +91,7 @@ public class MilestoneControllerIntegrationTest extends BaseControllerIntegratio
 
     @Test
     public void testCreateMilestones() throws Exception {
-        Competition newCompetition  = competitionRepository.save(newCompetition().withId(null).build());
+        Competition newCompetition  = competitionRepository.save(newCompetition().withId((Long)null).build());
 
         List<MilestoneResource> milestones = getMilestonesForCompetition(newCompetition.getId());
